지역 api 개발 로직 수정 - 카카오 api 사용

+++ src/main/java/com/exithere/rain/dto/request/DeviceJoinRequest.java
@@ -12,7 +12,7 @@ public class DeviceJoinRequest {
     private String deviceId;
     private String fcmToken;
     private String regionName;
-    private int regionX;
-    private int regionY;
+    private double regionX;
+    private double regionY;
 
 }
